pgdir_walk(): Always set PTE_W in page directory entries
If the first page of a page table is read-only, the page table's PDE will
be read-only too. The problem is that it won't be possible to create
read/write pages in that page table: since the PDE is read-only, all the
pages in that page-table are read-only by default!

Signed-off-by: Luiz Fernando N. Capitulino <lcapitulino@gmail.com>