page_fault_handler(): Simplify user page fault sanity checks
It's wrong to only check if the user has permissions to write in
the exception stack range.

The right thing to do is to check for write permissions at each time
we write in the user exception stack.

Also, we don't have to check the user page fault handler's pointer, it'll
fault if it's pointing inside the kernel.

This makes tests faultnostack, faultbadhandler and faultevilhandler
to succeed.

Signed-off-by: Luiz Fernando N. Capitulino <lcapitulino@gmail.com>