kernel: disable interrupts in _alltraps
If we don't disable interrupts in _alltraps the kernel will run
with interrupts enabled when the user issues a system call or
gets an exception. That's because traps and exceptions doesn't
disable interrupts automatically.

Note that it's not necessary to re-enable interrupts at the end
of _alltraps because iret will do that automatically (the user's
eflags has IF set).

Signed-off-by: Luiz Fernando N. Capitulino <lcapitulino@gmail.com>