fs: Fix block overwrite bug in file_get_block()
If file_get_block() is called for a dirty block which has
not synced with the disk yet, its data will be overwritten
by the in-disk version.

If the in-disk version is zeroed, for example, the in-memory
version will be cleaned.

That's a bug which doesn't allow one to created new files or
expand existing ones.

The fix is to check whether the block being read is already
mapped and dirty. If so, return it instead of reading the
block from disk again.

Signed-off-by: Luiz Fernando N. Capitulino <lcapitulino@gmail.com>