fs: read_block(): Kill PTE_D bit after block is read
If the PTE_D bit is not killed right after the block is read,
the block's page will *always* be written to disk even if the
client didn't call serve_dirty() for that page.

The PTE_D bit is killed by remapping the page with PTE_USER mask.

This change makes the file-system faster.

Signed-off-by: Luiz Fernando N. Capitulino <lcapitulino@gmail.com>