fs: file_get_block(): Fix block's page check
Commit 19a35a0dafb41d6fbee21440ac671eb25edc35a6 added a new
check in file_get_block(): if the block's page is dirty, return it
instead of reading the block from disk again.

Turns out that there are other strange page overwrite bugs which can
be triggered even with that check.

The right solution seems to be to check if the block's page is
_mapped_ - that's only half of the dirty block check - if the
block's page is mapped we do the same: return its in-memory address
instead of reading the block again.

I'm not sure whether this is the right fix, because I have the feeling
that soon or later the in-memory version of the block will became
out of sync with the in-disk version...

On the other hand, I couldn't prove that it's possible to happen and,
besides of fixing all the knows bugs so far, lab5 seems to state
that the check is indeed needed (see lab5's pdf p.7).

Signed-off-by: Luiz Fernando N. Capitulino <lcapitulino@gmail.com>